[AS] sharing state using spring and refactoring
